# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiError {
  error: string;
}

export interface LeadInput {
  name: string;
  email: string;
  phone: string;
  legalIssue: string;
  conversationId?: string;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  phone: string;
  legalIssue: string;
  conversationId?: string | null;
  createdAt: string;
}

export type VisitorTrackInputExtra = { [key: string]: unknown };

export interface VisitorTrackInput {
  sessionId: string;
  browser?: string;
  browserVersion?: string;
  os?: string;
  osVersion?: string;
  device?: string;
  screenWidth?: string;
  screenHeight?: string;
  language?: string;
  referrer?: string;
  pageUrl?: string;
  userAgent?: string;
  timezone?: string;
  extra?: VisitorTrackInputExtra;
}

export type VisitorSessionExtra = { [key: string]: unknown } | null;

export interface VisitorSession {
  id: number;
  sessionId: string;
  ip?: string | null;
  country?: string | null;
  region?: string | null;
  city?: string | null;
  timezone?: string | null;
  latitude?: string | null;
  longitude?: string | null;
  isp?: string | null;
  browser?: string | null;
  browserVersion?: string | null;
  os?: string | null;
  osVersion?: string | null;
  device?: string | null;
  screenWidth?: string | null;
  screenHeight?: string | null;
  language?: string | null;
  referrer?: string | null;
  pageUrl?: string | null;
  userAgent?: string | null;
  extra?: VisitorSessionExtra;
  visitedAt: string;
}