# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: submitLead
      tags: [leads]
      summary: Submit a new lead (contact form before chat)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /analytics/track:
    post:
      operationId: trackVisitor
      tags: [leads]
      summary: Track a visitor session with browser/device/location data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorTrackInput"
      responses:
        "201":
          description: Session tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorSession"
...
components:
  schemas:
...
    LeadInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        legalIssue:
          type: string
        conversationId:
          type: string
      required:
        - name
        - email
        - phone
        - legalIssue
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        legalIssue:
          type: string
        conversationId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - legalIssue
        - createdAt
    VisitorTrackInput:
      type: object
      properties:
        sessionId:
          type: string
        browser:
          type: string
        browserVersion:
          type: string
        os:
          type: string
        osVersion:
          type: string
        device:
          type: string
        screenWidth:
          type: string
        screenHeight:
          type: string
        language:
          type: string
        referrer:
          type: string
        pageUrl:
          type: string
        userAgent:
          type: string
        timezone:
          type: string
        extra:
          type: object
          additionalProperties: true
      required:
        - sessionId
    VisitorSession:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        ip:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        isp:
          type: string
          nullable: true
        browser:
          type: string
          nullable: true
        browserVersion:
          type: string
          nullable: true
        os:
          type: string
          nullable: true
        osVersion:
          type: string
          nullable: true
        device:
          type: string
          nullable: true
        screenWidth:
          type: string
          nullable: true
        screenHeight:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        referrer:
          type: string
          nullable: true
        pageUrl:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        extra:
          type: object
          additionalProperties: true
          nullable: true
        visitedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - visitedAt