# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadInput {
  name: string;
  email: string;
  phone: string;
  legalIssue: string;
  conversationId?: string;
}