# lib/api-zod/src/generated/types/visitorSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisitorSessionExtra } from './visitorSessionExtra';

export interface VisitorSession {
  id: number;
  sessionId: string;
  ip?: string | null;
  country?: string | null;
  region?: string | null;
  city?: string | null;
  timezone?: string | null;
  latitude?: string | null;
  longitude?: string | null;
  isp?: string | null;
  browser?: string | null;
  browserVersion?: string | null;
  os?: string | null;
  osVersion?: string | null;
  device?: string | null;
  screenWidth?: string | null;
  screenHeight?: string | null;
  language?: string | null;
  referrer?: string | null;
  pageUrl?: string | null;
  userAgent?: string | null;
  extra?: VisitorSessionExtra;
  visitedAt: Date;
}