# lib/api-zod/src/generated/types/visitorTrackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisitorTrackInputExtra } from './visitorTrackInputExtra';

export interface VisitorTrackInput {
  sessionId: string;
  browser?: string;
  browserVersion?: string;
  os?: string;
  osVersion?: string;
  device?: string;
  screenWidth?: string;
  screenHeight?: string;
  language?: string;
  referrer?: string;
  pageUrl?: string;
  userAgent?: string;
  timezone?: string;
  extra?: VisitorTrackInputExtra;
}